JAMES-3771 upgrade quota search es module to v8
