+++ mailbox/plugin/quota-search-elasticsearch-v8/src/test/resources/quotaRatioNoDomain.json
