+++ pom.xml
@@ -845,7 +845,7 @@
             </dependency>
             <dependency>
                 <groupId>${james.groupId}</groupId>
-                <artifactId>apache-james-mailbox-quota-search-elasticsearch-v7</artifactId>
+                <artifactId>apache-james-mailbox-quota-search-elasticsearch-v8</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
