+++ server/apps/cassandra-app/pom.xml
@@ -61,6 +61,11 @@
         <dependency>
             <groupId>${james.groupId}</groupId>
             <artifactId>apache-james-backends-es-v7</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>${james.groupId}</groupId>
+            <artifactId>apache-james-backends-es-v7</artifactId>
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
