+++ server/container/guice/elasticsearch/pom.xml
@@ -45,7 +45,7 @@
         </dependency>
         <dependency>
             <groupId>${james.groupId}</groupId>
-            <artifactId>apache-james-mailbox-quota-search-elasticsearch-v7</artifactId>
+            <artifactId>apache-james-mailbox-quota-search-elasticsearch-v8</artifactId>
         </dependency>
         <dependency>
             <groupId>${james.groupId}</groupId>
