+++ server/container/guice/elasticsearch/src/main/java/org/apache/james/modules/mailbox/ElasticSearchQuotaSearcherModule.java
@@ -28,19 +28,19 @@ import javax.inject.Inject;
 
 import org.apache.commons.configuration2.Configuration;
 import org.apache.commons.configuration2.ex.ConfigurationException;
-import org.apache.james.backends.es.v7.ElasticSearchConfiguration;
-import org.apache.james.backends.es.v7.ElasticSearchIndexer;
-import org.apache.james.backends.es.v7.ReactorElasticSearchClient;
+import org.apache.james.backends.es.v8.ElasticSearchConfiguration;
+import org.apache.james.backends.es.v8.ElasticSearchIndexer;
+import org.apache.james.backends.es.v8.ReactorElasticSearchClient;
 import org.apache.james.events.EventListener;
 import org.apache.james.lifecycle.api.Startable;
 import org.apache.james.mailbox.quota.QuotaRootResolver;
 import org.apache.james.quota.search.QuotaSearcher;
-import org.apache.james.quota.search.elasticsearch.v7.ElasticSearchQuotaConfiguration;
-import org.apache.james.quota.search.elasticsearch.v7.ElasticSearchQuotaSearcher;
-import org.apache.james.quota.search.elasticsearch.v7.QuotaSearchIndexCreationUtil;
-import org.apache.james.quota.search.elasticsearch.v7.UserRoutingKeyFactory;
-import org.apache.james.quota.search.elasticsearch.v7.events.ElasticSearchQuotaMailboxListener;
-import org.apache.james.quota.search.elasticsearch.v7.json.QuotaRatioToElasticSearchJson;
+import org.apache.james.quota.search.elasticsearch.v8.ElasticSearchQuotaConfiguration;
+import org.apache.james.quota.search.elasticsearch.v8.ElasticSearchQuotaSearcher;
+import org.apache.james.quota.search.elasticsearch.v8.QuotaSearchIndexCreationUtil;
+import org.apache.james.quota.search.elasticsearch.v8.UserRoutingKeyFactory;
+import org.apache.james.quota.search.elasticsearch.v8.events.ElasticSearchQuotaMailboxListener;
+import org.apache.james.quota.search.elasticsearch.v8.json.QuotaRatioToElasticSearchJson;
 import org.apache.james.utils.InitializationOperation;
 import org.apache.james.utils.InitilizationOperationBuilder;
 import org.apache.james.utils.PropertiesProvider;
