+++ server/protocols/webadmin/webadmin-mailbox/pom.xml
@@ -81,7 +81,7 @@
         </dependency>
         <dependency>
             <groupId>${james.groupId}</groupId>
-            <artifactId>apache-james-mailbox-quota-search-elasticsearch-v7</artifactId>
+            <artifactId>apache-james-mailbox-quota-search-elasticsearch-v8</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
